+++ R/weightit2ipt.R
@@ -67,7 +67,7 @@
 #'     }
 #'     \item{`quantile`}{a named list of quantiles (values between 0 and 1) for each continuous covariate, which are used to create additional variables that when balanced ensure balance on the corresponding quantile of the variable. For example, setting `quantile = list(x1 = c(.25, .5. , .75))` ensures the 25th, 50th, and 75th percentiles of `x1` in each treatment group will be balanced in the weighted sample. Can also be a single number (e.g., `.5`) or a vector (e.g., `c(.25, .5, .75)`) to request the same quantile(s) for all continuous covariates.
 #'     }
-#'   \item{`link`}{the link used to determine the inverse link for computing the (generalized) propensity scores. Default is `"logit"`, which is used in the original description of the method by Graham, Pinto, and Egel (2012), but `"probit"`, `"cauchit"`, `"cloglog"`, `"loglog"`, `"log"`, and `"clog"` are also allowed. Note that negative weights are possible with these last two and they should be used with caution. An object of class `"link-glm"` can also be supplied. The argument is passed to [quasibinomial()].
+#'   \item{`link`}{the link used to determine the inverse link for computing the (generalized) propensity scores. Default is `"logit"`, which is used in the original description of the method by Graham, Pinto, and Egel (2012), but `"probit"`, `"cauchit"`, `"cloglog"`, `"loglog"`, `"log"`, `"clog"`, and `"softplus"` are also allowed. Note that negative weights are possible with these last two and they should be used with caution. An object of class `"link-glm"` can also be supplied. The argument is passed to [quasibinomial()].
 #'   }
 #' }
 #'
@@ -165,7 +165,7 @@ weightit2ipt <- function(covs, treat, s.weights, subset, estimand, focal,
 
   if (is.character(link) && length(link) == 1L) {
     arg::arg_element(link, c("logit", "probit", "cloglog", "loglog",
-                             "cauchit", "log", "clog"))
+                             "cauchit", "log", "clog", "softplus"))
 
     link <- .make_link(link)
   }
@@ -381,7 +381,8 @@ weightit2ipt.multi <- function(covs, treat, s.weights, subset, estimand, focal,
   link <- ...get("link", "logit")
 
   if (is.character(link) && length(link) == 1L) {
-    arg::arg_element(link, c("logit", "probit", "cloglog", "loglog", "cauchit", "log", "clog"))
+    arg::arg_element(link, c("logit", "probit", "cloglog", "loglog",
+                             "cauchit", "log", "clog", "softplus"))
 
     link <- .make_link(link)
   }